backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pdf2image==1.16.3
Pillow==10.1.0
pypdf>=3.0.0
PyPDF2==3.0.1
pdfplumber==0.10.3
opencv-python==4.8.1.78
sentence-transformers>=2.7.0
openai==1.3.5
numpy>=1.26.0,<2.0.0
scikit-learn>=1.3.2
pydantic==2.5.0
aiofiles==23.2.1
python-dotenv==1.0.0

